Difference equation

RegionalGrasslandSim.initial_conditions — Method

initial_conditions(; initbiomass, npatches, nspecies)

Initialize all state variables with the same biomass and water content.

RegionalGrasslandSim.initialize_parameters — Method

initialize_parameters(; input_obj)

Initialize all parameters and store them in one object.

RegionalGrasslandSim.main_loop! — Method

main_loop!(; calc, ts, p)

Run the main loop for all days.

RegionalGrasslandSim.one_day! — Method

one_day!(; calc, p, t)

Calculate the density differences of all state variables of one day.

RegionalGrasslandSim.solve_prob — Method

solve_prob(; input_obj)

Solve the model for one site.

Intialize the parameters, the state variables and the output vectors. In addition some vectors are preallocated to avoid allocations in the main loop. Then, run the main loop and store the results in the output vectors.